requirements_314_full.txt
absl-py==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.77.1
anyio==4.12.1
async-generator==1.10
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.15
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
edge-tts==7.2.7
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
fpdf2==2.8.5
frozenlist==1.8.0
google-auth==2.48.0
google-genai==1.61.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
keyboard==0.13.5
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.32
MouseInfo==0.1.3
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.2
openai==2.16.0
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.0
pillow==12.1.0
propcache==0.4.1
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pynput==1.8.1
pyparsing==3.3.2
pypdf==6.6.2
pyperclip==1.11.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pyttsx3==2.99
pytweening==1.2.0
pyvts==0.3.3
pywin32==311
requests==2.32.5
rsa==4.9.1
scipy==1.17.0
selenium==4.40.0
setuptools==80.10.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.5
soupsieve==2.8.3
SpeechRecognition==3.14.5
standard-aifc==3.13.0
standard-chunk==3.13.0
tabulate==0.9.0
tenacity==9.1.2
tqdm==4.67.3
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.5
wheel==0.46.3
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
